requirements.txt
Flask==2.2.5
pymongo==4.5.0
Werkzeug==2.2.3
gunicorn==21.2.0
python-dotenv==0.21.1
PyJWT==2.8.0
yagmail==0.15.293
